feat: 项目结构调整

+++ strr-api/strr-api-system/src/main/java/com/strr/system/api/RemoteUserService.java
@@ -16,5 +16,5 @@ public interface RemoteUserService {
     /**
      * 获取用户权限
      */
-    List<String> listPerms(Integer userId);
+    List<String> listPerms(Long userId);
 }
