+++ strr-module/strr-data/src/main/java/com/strr/data/mapper/DmsColumnMapper.java
@@ -8,7 +8,7 @@ import java.util.List;
 /**
  * 业务字段信息
  */
-public interface DmsColumnMapper extends CrudMapper<DmsColumn, Integer> {
+public interface DmsColumnMapper extends CrudMapper<DmsColumn, Long> {
     /**
      * 根据表名获取数据库字段信息
      */
@@ -17,5 +17,5 @@ public interface DmsColumnMapper extends CrudMapper<DmsColumn, Integer> {
     /**
      * 根据 tableId 删除字段信息
      */
-    void removeByTableId(Integer tableId);
+    void removeByTableId(Long tableId);
 }
