+++ strr-module/strr-data/src/main/java/com/strr/data/model/DmsTable.java
@@ -1,8 +1,8 @@
 package com.strr.data.model;
 
-import com.strr.base.annotation.Id;
-import com.strr.base.annotation.Table;
 import com.strr.base.model.BaseModel;
+import com.strr.mybatis.annotation.Id;
+import com.strr.mybatis.annotation.Table;
 
 /**
  * 业务表信息
@@ -13,7 +13,7 @@ public class DmsTable extends BaseModel {
      * 编号
      */
     @Id
-    private Integer id;
+    private Long id;
 
     /**
      * 表名称
@@ -35,11 +35,11 @@ public class DmsTable extends BaseModel {
      */
     private String status;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
