+++ strr-module/strr-data/src/main/java/com/strr/data/model/vo/DmsModuleVo.java
@@ -12,12 +12,12 @@ public class DmsModuleVo {
     /**
      * 编号
      */
-    private Integer id;
+    private Long id;
 
     /**
      * 归属表编号
      */
-    private Integer tableId;
+    private Long tableId;
 
     /**
      * 模块名称
@@ -54,19 +54,19 @@ public class DmsModuleVo {
      */
     private List<DmsColumn> columns;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getTableId() {
+    public Long getTableId() {
         return tableId;
     }
 
-    public void setTableId(Integer tableId) {
+    public void setTableId(Long tableId) {
         this.tableId = tableId;
     }
