+++ strr-module/strr-resource/src/main/java/com/strr/resource/mapper/SysOssConfigMapper.java
@@ -3,5 +3,5 @@ package com.strr.resource.mapper;
 import com.strr.base.mapper.CrudMapper;
 import com.strr.resource.model.SysOssConfig;
 
-public interface SysOssConfigMapper extends CrudMapper<SysOssConfig, Integer> {
+public interface SysOssConfigMapper extends CrudMapper<SysOssConfig, Long> {
 }
