+++ strr-module/strr-resource/src/main/java/com/strr/resource/mapper/SysOssMapper.java
@@ -3,5 +3,5 @@ package com.strr.resource.mapper;
 import com.strr.base.mapper.CrudMapper;
 import com.strr.resource.model.SysOss;
 
-public interface SysOssMapper extends CrudMapper<SysOss, Integer> {
+public interface SysOssMapper extends CrudMapper<SysOss, Long> {
 }
