+++ strr-module/strr-resource/src/main/java/com/strr/resource/model/SysOss.java
@@ -1,8 +1,8 @@
 package com.strr.resource.model;
 
-import com.strr.base.annotation.Id;
-import com.strr.base.annotation.Table;
 import com.strr.base.model.BaseModel;
+import com.strr.mybatis.annotation.Id;
+import com.strr.mybatis.annotation.Table;
 
 /**
  * 对象存储
@@ -13,7 +13,7 @@ public class SysOss extends BaseModel {
      * 主键
      */
     @Id
-    private Integer id;
+    private Long id;
 
     /**
      * 服务商
@@ -40,11 +40,11 @@ public class SysOss extends BaseModel {
      */
     private String url;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
