+++ strr-module/strr-resource/src/main/java/com/strr/resource/service/impl/SysOssConfigServiceImpl.java
@@ -1,7 +1,7 @@
 package com.strr.resource.service.impl;
 
 import com.strr.base.service.impl.CrudServiceImpl;
-import com.strr.constant.Constant;
+import com.strr.base.constant.Constant;
 import com.strr.resource.mapper.SysOssConfigMapper;
 import com.strr.resource.model.SysOssConfig;
 import com.strr.resource.service.ISysOssConfigService;
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 
 @Service
-public class SysOssConfigServiceImpl extends CrudServiceImpl<SysOssConfig, Integer> implements ISysOssConfigService {
+public class SysOssConfigServiceImpl extends CrudServiceImpl<SysOssConfig, Long> implements ISysOssConfigService {
     private final SysOssConfigMapper sysOssConfigMapper;
 
     public SysOssConfigServiceImpl(SysOssConfigMapper sysOssConfigMapper) {
