+++ strr-module/strr-system/src/main/java/com/strr/system/mapper/SysDictDataMapper.java
@@ -3,5 +3,5 @@ package com.strr.system.mapper;
 import com.strr.base.mapper.CrudMapper;
 import com.strr.system.model.SysDictData;
 
-public interface SysDictDataMapper extends CrudMapper<SysDictData, Integer> {
+public interface SysDictDataMapper extends CrudMapper<SysDictData, Long> {
 }
