+++ strr-module/strr-system/src/main/java/com/strr/system/mapper/SysDictTypeMapper.java
@@ -3,5 +3,5 @@ package com.strr.system.mapper;
 import com.strr.base.mapper.CrudMapper;
 import com.strr.system.model.SysDictType;
 
-public interface SysDictTypeMapper extends CrudMapper<SysDictType, Integer> {
+public interface SysDictTypeMapper extends CrudMapper<SysDictType, Long> {
 }
