+++ strr-module/strr-system/src/main/java/com/strr/system/mapper/SysResourceMapper.java
@@ -6,9 +6,9 @@ import org.apache.ibatis.annotations.Param;
 
 import java.util.List;
 
-public interface SysResourceMapper extends CrudMapper<SysResource, Integer> {
+public interface SysResourceMapper extends CrudMapper<SysResource, Long> {
     /**
      * 获取用户权限
      */
-    List<SysResource> listByUserId(@Param("userId") Integer userId, @Param("menu") String menu);
+    List<SysResource> listByUserId(@Param("userId") Long userId, @Param("menu") String menu);
 }
