+++ strr-module/strr-system/src/main/java/com/strr/system/mapper/SysRoleMapper.java
@@ -3,5 +3,5 @@ package com.strr.system.mapper;
 import com.strr.system.model.SysRole;
 import com.strr.base.mapper.CrudMapper;
 
-public interface SysRoleMapper extends CrudMapper<SysRole, Integer> {
+public interface SysRoleMapper extends CrudMapper<SysRole, Long> {
 }
