+++ strr-module/strr-system/src/main/java/com/strr/system/mapper/SysUserMapper.java
@@ -3,7 +3,7 @@ package com.strr.system.mapper;
 import com.strr.system.model.SysUser;
 import com.strr.base.mapper.CrudMapper;
 
-public interface SysUserMapper extends CrudMapper<SysUser, Integer> {
+public interface SysUserMapper extends CrudMapper<SysUser, Long> {
     /**
      * 获取用户信息
      */
